restructure ioreport into typed modules
